requirements.txt
flask==3.0.0
flask-cors==4.0.0
langchain==0.1.0
langchain-openai==0.0.5
langchain-community==0.0.10
pydantic==2.5.0
python-dotenv==1.0.0
supabase==2.0.0
spotipy==2.22.1